Squashed commit of the following:
commit 1d0c34c818be7bf640eb874fa28079cf80088634
Author: John Chapman <contact@john-chapman.net>
Date:   Sat Aug 26 15:23:56 2017 +0200

    Profiler: Marker tracking, improved UI

commit 377292ea632aaf88f15df28a42816d05575d510c
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 25 09:31:33 2017 +0200

    Profiler: Marker tracking

commit f976f7088e52185ddb1c2b0a82a5d47479fe1348
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 25 09:31:17 2017 +0200

    AppSample: Removed properties shortcut

commit 1b23ce754ce6eb4211045719e2b35a4d047b7e5f
Author: John Chapman <contact@john-chapman.net>
Date:   Tue Aug 22 09:30:00 2017 +0200

    Update ImGui, set default color edit options, profiler viewer WIP

commit 1df862dc20d3b6aecbc811cc6b81cf60019fe68c
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 18 16:58:01 2017 +0200

    Profiler: Viewer WIP (values)

commit 3de982002a0724f668daa1d4996aa0e3c5a7c1d2
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 18 16:57:18 2017 +0200

    Merge master

commit 417ff08a665c18a893c9a62adc124106b32464a4
Merge: ab9d454 3ded5bb
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 18 16:19:41 2017 +0200

    Merge branch 'master' into profiler_tracker

commit ab9d45407959f2437c1986c9633e42b1df05da57
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 11 09:13:29 2017 +0200

    Profiler: Disable hover behavior if window not focussed

commit ef75813f35b7f9ebd79525ac954838c167ccf385
Merge: 1b9614f 085d8ab
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 11 09:08:50 2017 +0200

    Merge branch 'master' into profiler_tracker

commit 1b9614fc099ce5838622702ce07c4f74b2f37297
Author: John Chapman <contact@john-chapman.net>
Date:   Fri Aug 11 08:45:05 2017 +0200

    Profiler: Minor usability improvements

commit da3c64461347691d067c38f9ace534494b53aea2
Author: John <j.x.chapman@gmail.com>
Date:   Thu Aug 10 18:44:00 2017 +0200

    WIP

+++ src/all/frm/AppSample.h
@@ -43,7 +43,7 @@ public:
 	const Framebuffer*  getDefaultFramebuffer() const                 { return m_fbDefault; }
 	void                setDefaultFramebuffer(const Framebuffer* _fb) { m_fbDefault = _fb; }
 
-	uint                getFrameIndex() const         { return m_frameIndex; }
+	uint64              getFrameIndex() const         { return m_frameIndex; }
 
 protected:		
 	typedef apt::FileSystem::PathStr PathStr;
